cr fix spelling, move cliapp to subpkg of internal pkg
